Condense special distributions to one file
