+++ alea/discrete/__init__.py
@@ -1,4 +1,2 @@
 from .discrete_randvar import *
-from .uniform_randvar import *
-from .bernoulli_randvar import *
-from .binomial_randvar import *
+from .special_randvar import *
